FDF with perspective

+++ fdf_new/draw.c
@@ -46,7 +46,7 @@ int			draw_horiz(ft_vector v1, ft_vector v2, void *mlx_ptr, void *mlx_win)
 	dir = -1;
 	if (x < v2.x)
 		dir = 1;
-	while (x != v2.x)
+	while ((int)x != (int)v2.x)
 	{
 		mlx_pixel_put(mlx_ptr, mlx_win, x, y, v1.c);
 		x += dir;
@@ -70,7 +70,7 @@ int			draw_vert(ft_vector v1, ft_vector v2, void *mlx_ptr, void *mlx_win)
 	dir = -1;
 	if (y < v2.y)
 		dir = 1;
-	while (y != v2.y)
+	while ((int)y != (int)v2.y)
 	{
 		mlx_pixel_put(mlx_ptr, mlx_win, x, y, v1.c);
 		y += dir;
@@ -102,7 +102,7 @@ int			draw_line(ft_vector v1, ft_vector v2, void *mlx_ptr, void *mlx_win)
 	ydif = -1;
 	if ((v2.y - v1.y) > 0)
 		ydif = 1;
-	while (x != v2.x && y != v2.y)
+	while ((int)x != (int)v2.x && (int)y != (int)v2.y)
 	{
 		mlx_pixel_put(mlx_ptr, mlx_win, x, y, v1.c);
 		if (m > 1)
